+++ tools/build.py
@@ -10,6 +10,7 @@ official CI/CD build implementation in place.
 
 Changelog:
 
+0.4.18 [24.04.19] CEP plugin support. PySide integrations not platform dependent.
 0.4.17 [24.04.18] Build script to support extras.
 0.4.16 [24.03.19] PySide6 resource build support.
 0.4.15 [24.03.13] Fix platform dependent bug.
@@ -45,7 +46,7 @@ from distutils.spawn import find_executable
 import fileinput
 import tempfile
 
-__version__ = '0.4.17'
+__version__ = '0.4.19'
 
 ZXPSIGN_CMD = 'ZXPSignCmd'
 
@@ -70,7 +71,6 @@ def build_package(pkg_path, args, command=None):
     CONNECT_PLUGIN_PATH = os.path.join(ROOT_PATH, 'connect-plugin')
     BUILD_PATH = os.path.join(ROOT_PATH, 'dist')
     EXTENSION_PATH = os.path.join(ROOT_PATH, 'extensions')
-    CEP_PATH = os.path.join(ROOT_PATH, 'resource', 'cep')
     USES_FRAMEWORK = False
     FTRACK_DEP_LIBS = {}
     PLATFORM_DEPENDENT = False
@@ -136,10 +136,7 @@ def build_package(pkg_path, args, command=None):
                         VERSION = line.split('=')[1].strip().strip('"')
                 elif section == 'tool.poetry.dependencies':
                     if line.find('pyside') > -1:
-                        PLATFORM_DEPENDENT = True
-                        logging.info(
-                            'Platform dependent build - OS suffix will be added to artifact.'
-                        )
+                        pass
 
         append_dependencies(POETRY_CONFIG_PATH)
 
@@ -187,11 +184,7 @@ def build_package(pkg_path, args, command=None):
         )
         with open(source_path, 'r') as f_src:
             with open(target_path, 'w') as f_dst:
-                f_dst.write(
-                    f_src.read().replace(
-                        '{{FTRACK_FRAMEWORK_PHOTOSHOP_VERSION}}', VERSION
-                    )
-                )
+                f_dst.write(f_src.read().replace('${VERSION}', VERSION))
 
     def build_connect_plugin(args):
         '''
@@ -579,12 +572,7 @@ def build_package(pkg_path, args, command=None):
         Qt.
 
         '''
-        replace = r'''
-try: 
-    from PySide6 import QtCore
-except ImportError: 
-    from PySide2 import QtCore
-'''
+        replace = r'from Qt import QtCore'
         for line in fileinput.input(
             resource_target_path, inplace=True, mode='r'
         ):
@@ -637,10 +625,7 @@ except ImportError:
         else:
             logging.warning('No styles to compile.')
 
-        pyside_version = args.pyside_version
-        if not pyside_version:
-            pyside_version = "2"
-        pyside_rcc_command = f'pyside{pyside_version}-rcc'
+        pyside_rcc_command = 'pyside2-rcc'
         try:
             executable = None
 
@@ -688,10 +673,15 @@ except ImportError:
     def build_cep(args):
         '''Wrapper for building Adobe CEP extension'''
 
+        CEP_PATH = os.path.join(MONOREPO_PATH, 'resource', 'adobe-cep')
         if not os.path.exists(CEP_PATH):
-            raise Exception('Missing "{}/" folder!'.format(CEP_PATH))
+            raise Exception('Missing common "{}/" folder!'.format(CEP_PATH))
 
-        MANIFEST_PATH = os.path.join(CEP_PATH, 'bundle', 'manifest.xml')
+        CEP_DCC_PATH = os.path.join(ROOT_PATH, 'resource', 'adobe-cep')
+        if not os.path.exists(CEP_DCC_PATH):
+            raise Exception('Missing DCC "{}/" folder!'.format(CEP_PATH))
+
+        MANIFEST_PATH = os.path.join(CEP_DCC_PATH, 'bundle', 'manifest.xml')
         if not os.path.exists(MANIFEST_PATH):
             raise Exception('Missing manifest:{}!'.format(MANIFEST_PATH))
 
@@ -713,6 +703,7 @@ except ImportError:
             )
 
         STAGING_PATH = os.path.join(BUILD_PATH, 'staging')
+        assert DCC_NAME, 'Please provide DCC name to build CEP plugin for'
 
         # Clean previous build
         if os.path.exists(BUILD_PATH):
@@ -744,8 +735,8 @@ except ImportError:
         if not os.path.exists(style_path):
             raise Exception('Missing "{}/" folder!'.format(style_path))
 
-        # Copy html
-        for filename in ['index.html']:
+        # Copy html and base bootstrap
+        for filename in ['index.html', 'bootstrap.js']:
             parse_and_copy(
                 os.path.join(CEP_PATH, filename),
                 os.path.join(STAGING_PATH, filename),
@@ -788,26 +779,20 @@ except ImportError:
         )
 
         logging.info("Copying framework js lib files")
+        framework_js_path = os.path.join(
+            MONOREPO_PATH, 'libs', f'framework-js', 'source'
+        )
         for js_file in [
             os.path.join(
-                MONOREPO_PATH,
-                'projects',
-                'framework-photoshop-js',
-                'source',
+                framework_js_path,
                 'utils.js',
             ),
             os.path.join(
-                MONOREPO_PATH,
-                'projects',
-                'framework-photoshop-js',
-                'source',
+                framework_js_path,
                 'event-constants.js',
             ),
             os.path.join(
-                MONOREPO_PATH,
-                'projects',
-                'framework-photoshop-js',
-                'source',
+                framework_js_path,
                 'events-core.js',
             ),
         ]:
@@ -815,13 +800,22 @@ except ImportError:
                 js_file,
                 os.path.join(STAGING_PATH, 'lib', os.path.basename(js_file)),
             )
-        for filename in ['bootstrap.js', 'ps.jsx']:
+
+        # Copy extensions
+        if DCC_NAME == 'photoshop':
+            extendscript_file = 'ps{}.jsx'
+        else:
+            raise Exception('Unsupported Adobe DCC: {}'.format(DCC_NAME))
+
+        for filename in [
+            'bootstrap-dcc.js',
+            extendscript_file.format(''),
+            extendscript_file.format('-include'),
+        ]:
             parse_and_copy(
                 os.path.join(
-                    MONOREPO_PATH,
-                    'projects',
-                    'framework-photoshop-js',
-                    'source',
+                    EXTENSION_PATH,
+                    'js',
                     filename,
                 ),
                 os.path.join(STAGING_PATH, filename),
@@ -836,8 +830,7 @@ except ImportError:
         parse_and_copy(MANIFEST_PATH, manifest_staging_path)
 
         extension_output_path = os.path.join(
-            BUILD_PATH,
-            'ftrack-framework-adobe-{}.zxp'.format(VERSION),
+            BUILD_PATH, f'ftrack-framework-{DCC_NAME}-{VERSION}.zxp'
         )
 
         if not args.nosign:
@@ -997,16 +990,10 @@ if __name__ == '__main__':
         help='(QT resource build) Override the default style path (resource/style).',
     )
     parser.add_argument(
-        '--pyside_version',
-        help='(QT resource build) The version of PySide to use, default is "2"',
-    )
-    # QT/RV shared
-    parser.add_argument(
         '--output_path',
         help='(QT resource build/RV pkg build) Override the QT resource output directory.',
     )
 
-    # CEP options
     parser.add_argument(
         '--nosign',
         help='(CEP plugin build) Do not sign and create ZXP.',