+++ projects/framework-photoshop/release_notes.md
@@ -2,6 +2,8 @@
 
 ## Upcoming
 
+* [changed] Moved extendscript jsx resources to extensions js folder, added PS specific bootstrap extension.
+* [changed] Refactored the code to enable code sharing with other Adobe CEP integrations.
 * [changed] Replace Qt.py imports to PySide2 and PySide6 on widgets.
 
 ## v24.4.1